Create a separate class for handling the analyzable files
